+++ src/App.js
@@ -3,11 +3,10 @@ import { Routes, Route } from "react-router-dom"
 import PaymentForm from './components/Forms';
 import Home from './components/Home';
 
-import LiveRecord, { save as saveData} from './utils/store';
+import { JSONtoString, collection, database, save as saveData, strorageKey} from './utils/store';
+import { onValue, ref } from 'firebase/database';
 import './App.css';
 
-// Listen for update.
-LiveRecord()
 
 function App() {
 
@@ -22,3 +21,22 @@ function App() {
 }
 
 export default App;
+
+
+const paymentRef = ref(database, collection);
+
+onValue(paymentRef, (snapshot) => {
+  const data = snapshot.val() || {};
+  const entry = {};
+
+  Object.entries(data).forEach(([remoteId, remoteData]) => {
+    const { payId, ...rest } = remoteData;
+
+    entry[payId] = {
+      id: remoteId,
+      ...rest
+    }
+  })
+
+  localStorage.setItem(strorageKey, JSONtoString(entry));
+});
